requirements_local.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.18
Django==1.8.1
djangorestframework==3.3.3
django-autoslug==1.9.3
django-debug-toolbar==1.3.0
django-filter==0.13.0
gunicorn==19.3.0
kombu==3.0.35
Markdown==2.6.6
pytz==2016.3
redis==2.10.3
requests==2.9.1
sqlparse==0.1.19
whitenoise==1.0.6
